Make PageTriage wiki-agnostic
Creates the config variable $wgPageTriageEnableEnglishWikipediaFeatures. Sets its default to true to keep default behavior the same (default behavior is to be enwiki-specific). Can be set to false for non-enwikis.

The implementation is the simplest possible: some conditionals to hide HTML for parts of PageTriage that are enwiki-specific. When $wgPageTriageEnableEnglishWikipediaFeatures == false, hides the following:

- Special:NewPagesFeed -> Filters -> "Nominated for deletion"
- Special:NewPagesFeed -> Filters -> "Predicted class: stub, start, c-class, b-class, good, featured"
- Page Curation toolbar -> Maintenance tagging menu
- Page Curation toolbar -> Deletion tagging menu
- Also turns off the pagetriagetagging API.

Bug: T50552

+++ modules/ext.pageTriage.views.toolbar/ToolbarView.js
@@ -52,13 +52,13 @@ ToolbarView = Backbone.View.extend( {
 		}
 
 		// add tags
-		if ( this.isFlyoutEnabled( 'tags' ) ) {
+		if ( this.isFlyoutEnabled( 'tags' ) && config.PageTriageEnableEnglishWikipediaFeatures ) {
 			var TagsView = require( './tags.js' );
 			tools.push( new TagsView( { eventBus: eventBus, model: article, moduleConfig: modules.tags } ) );
 		}
 
 		// delete
-		if ( this.isFlyoutEnabled( 'delete' ) ) {
+		if ( this.isFlyoutEnabled( 'delete' ) && config.PageTriageEnableEnglishWikipediaFeatures ) {
 			var DeleteView = require( './delete.js' );
 			tools.push( new DeleteView( { eventBus: eventBus, model: article, moduleConfig: modules.delete } ) );
 		}
